🚚 api: Move v3-certification-course-details-for-administration serializer to session-management context
