+++ api/src/certification/session-management/infrastructure/serializers/v3-certification-course-details-for-administration-serializer.js
